# service/api-gateway/docs/swagger/swagger.yaml
basePath: /
definitions:
  model.EmailRegister:
    properties:
      email:
        example: idea_wj@163.com
        type: string
      password:
        example: sja123
        type: string
    type: object
  model.Kline:
    properties:
      apiKey:
        example: cf998e2f-ed32b3f8-42c02442-1011e
        type: string
      apiSecret:
        example: 91e6e532-3bb37cad-c0a689ca-75ee5
        type: string
      currencyPair:
        example: BTC_USDT
        type: string
      exName:
        example: huobi.pro
        type: string
      period:
        example: 0
        type: integer
      since:
        example: 10000000
        type: integer
      size:
        example: 150
        type: integer
    type: object
  model.OrderBase:
    properties:
      apiKey:
        example: cf998e2f-ed32b3f8-42c02442-1011e
        type: string
      apiSecret:
        example: 91e6e532-3bb37cad-c0a689ca-75ee5
        type: string
      currencyPair:
        example: BTC_USDT
        type: string
      exName:
        example: huobi.pro
        type: string
      orderId:
        example: 2xxxxx
        type: string
    type: object
  model.OrderBase2:
    properties:
      apiKey:
        example: cf998e2f-ed32b3f8-42c02442-1011e
        type: string
      apiSecret:
        example: 91e6e532-3bb37cad-c0a689ca-75ee5
        type: string
      currencyPair:
        example: BTC_USDT
        type: string
      exName:
        example: huobi.pro
        type: string
    type: object
  model.OrderBase3:
    properties:
      apiKey:
        example: cf998e2f-ed32b3f8-42c02442-1011e
        type: string
      apiSecret:
        example: 91e6e532-3bb37cad-c0a689ca-75ee5
        type: string
      currencyPair:
        example: BTC_USDT
        type: string
      exName:
        example: huobi.pro
        type: string
      size:
        example: 1
        type: integer
    type: object
  model.OrderBase4:
    properties:
      apiKey:
        example: cf998e2f-ed32b3f8-42c02442-1011e
        type: string
      apiSecret:
        example: 91e6e532-3bb37cad-c0a689ca-75ee5
        type: string
      currencyPair:
        example: BTC_USDT
        type: string
      exName:
        example: huobi.pro
        type: string
      since:
        example: 1
        type: integer
    type: object
  model.OrderHistory:
    properties:
      apiKey:
        example: cf998e2f-ed32b3f8-42c02442-1011e
        type: string
      apiSecret:
        example: 91e6e532-3bb37cad-c0a689ca-75ee5
        type: string
      currencyPair:
        example: BTC_USDT
        type: string
      currentPage:
        example: 1
        type: integer
      exName:
        example: huobi.pro
        type: string
      pageSize:
        example: 1
        type: integer
    type: object
  model.OrderPlace:
    properties:
      accountType:
        example: 2
        type: integer
      amount:
        example: "0"
        type: string
      apiKey:
        example: cf998e2f-ed32b3f8-42c02442-1011e
        type: string
      apiSecret:
        example: 91e6e532-3bb37cad-c0a689ca-75ee5
        type: string
      currencyPair:
        example: BTC_USDT
        type: string
      exName:
        example: huobi.pro
        type: string
      orderType:
        example: 2
        type: integer
      price:
        example: "0"
        type: string
    type: object
  model.TokenAuth:
    properties:
      apiKey:
        example: cf998e2f-ed32b3f8-42c02442-1011e
        type: string
      apiSecret:
        example: 91e6e532-3bb37cad-c0a689ca-75ee5
        type: string
      exName:
        example: huobi.pro
        type: string
    type: object
host: localhost:8090
info:
  contact:
    email: idea_wj
    name: idea_wj@163.com
    url: http://www.wjblog.com/
  description: 暂无描述
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://www.wjblog.com/
  title: bitesla
  version: "2.0"
paths:
  /trader/cancelOrder:
    post:
      consumes:
      - application/json
      description: 此接口发送一个撤销订单的请求
      parameters:
      - description: 每个参数均不得为空
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.OrderBase'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功与否
          schema:
            type: string
      summary: 此接口发送一个撤销订单的请求
      tags:
      - 交易所相关
  /trader/getAccount:
    post:
      consumes:
      - application/json
      description: 获取当前账户信息
      parameters:
      - description: 每个参数均不得为空
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.TokenAuth'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功与否
          schema:
            type: string
      summary: 获取当前账户信息
      tags:
      - 交易所相关
  /trader/getDepth:
    post:
      consumes:
      - application/json
      description: 返回指定交易对的当前市场深度数据
      parameters:
      - description: 每个参数均不得为空
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.OrderBase3'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功与否
          schema:
            type: string
      summary: 返回指定交易对的当前市场深度数据
      tags:
      - 交易所相关
  /trader/getExchangeName:
    post:
      consumes:
      - application/json
      description: 获取交易所名称
      parameters:
      - description: 每个参数均不得为空
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.TokenAuth'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功与否
          schema:
            type: string
      summary: 获取交易所名称
      tags:
      - 交易所相关
  /trader/getKlineRecords:
    post:
      consumes:
      - application/json
      description: 获取k线
      parameters:
      - description: 每个参数均不得为空
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.Kline'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功与否
          schema:
            type: string
      summary: 获取k线
      tags:
      - 交易所相关
  /trader/getOneOrder:
    post:
      consumes:
      - application/json
      description: 返回订单的最新状况和详情
      parameters:
      - description: 每个参数均不得为空
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.OrderBase'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功与否
          schema:
            type: string
      summary: 返回订单的最新状况和详情
      tags:
      - 交易所相关
  /trader/getOrderHistorys:
    post:
      consumes:
      - application/json
      description: 获取某交易对交易历史
      parameters:
      - description: 每个参数均不得为空
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.OrderHistory'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功与否
          schema:
            type: string
      summary: 获取某交易对交易历史
      tags:
      - 交易所相关
  /trader/getTicker:
    post:
      consumes:
      - application/json
      description: 获取ticker信息，同时提供24小时交易聚合信息
      parameters:
      - description: 每个参数均不得为空
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.OrderBase2'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功与否
          schema:
            type: string
      summary: 获取ticker信息，同时提供24小时交易聚合信息
      tags:
      - 交易所相关
  /trader/getTrades:
    post:
      consumes:
      - application/json
      description: 非个人，整个交易所的交易记录
      parameters:
      - description: 每个参数均不得为空
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.OrderBase4'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功与否
          schema:
            type: string
      summary: 非个人，整个交易所的交易记录
      tags:
      - 交易所相关
  /trader/getUnfinishOrders:
    post:
      consumes:
      - application/json
      description: 未完成的交易订单
      parameters:
      - description: 每个参数均不得为空
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.OrderBase2'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功与否
          schema:
            type: string
      summary: 未完成的交易订单
      tags:
      - 交易所相关
  /trader/orderPlace:
    post:
      consumes:
      - application/json
      description: 发送一个新的订单到某交易所进行撮合
      parameters:
      - description: "每个参数均不得为空,//\tOrderType: 0 表示limitBuy 1 表示limitSell 2 表示marketBuy
          3 表示marketSell ; AccountType: 1 表示point 2 表示splot"
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.OrderPlace'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功与否
          schema:
            type: string
      summary: 发送一个新的订单到某交易所进行撮合
      tags:
      - 交易所相关
  /user/registerEmail:
    post:
      consumes:
      - application/json
      description: 邮箱注册
      parameters:
      - description: 每个参数均不得为空
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/model.EmailRegister'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功与否
          schema:
            type: string
      summary: 邮箱注册
      tags:
      - 账户操作
securityDefinitions:
  token:
    in: header
    name: token
    type: apiKey
swagger: "2.0"